+++ src/main/java/chess/domain/game/PiecePositionInitializer.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import chess.domain.position.BoardPosition;
 import chess.domain.position.Lettering;
