+++ src/main/java/chess/domain/piece/movestrategy/MovablePositionGenerator.java
@@ -1,6 +1,6 @@
 package chess.domain.piece.movestrategy;
 
-import chess.domain.PiecePosition;
+import chess.domain.game.PiecePosition;
 import chess.domain.piece.Piece;
 import chess.domain.position.BoardPosition;
 import chess.domain.position.Position;
