+++ src/test/java/chess/domain/game/ChessStatusTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import static chess.domain.TestSetting.A1;
 import static chess.domain.TestSetting.A2;
