+++ src/test/java/chess/domain/game/PiecePositionInitializerTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
