+++ src/test/java/chess/domain/game/TurnExecutorTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import static chess.domain.TestSetting.D2;
 import static chess.domain.TestSetting.D3;
