+++ src/test/java/chess/domain/piece/movestrategy/KingMoveStrategyTest.java
@@ -13,7 +13,7 @@ import static chess.domain.TestSetting.KING_BLACK;
 import static chess.domain.TestSetting.KING_WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.PiecePosition;
+import chess.domain.game.PiecePosition;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 import java.util.HashMap;
