+++ src/test/java/chess/domain/piece/movestrategy/PawnMoveStrategyTest.java
@@ -16,7 +16,7 @@ import static chess.domain.TestSetting.PAWN_WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.domain.PiecePosition;
+import chess.domain.game.PiecePosition;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 import java.util.HashMap;
