+++ src/test/java/chess/domain/piece/movestrategy/QueenMoveStrategyTest.java
@@ -32,7 +32,7 @@ import static chess.domain.TestSetting.QUEEN_BLACK;
 import static chess.domain.TestSetting.QUEEN_WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.PiecePosition;
+import chess.domain.game.PiecePosition;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 import java.util.HashMap;
